requirements.txt
# Production Dependencies
# Core Data Science
numpy>=1.26.0,<2.0.0
pandas>=2.1.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
xgboost>=1.7.0,<3.0.0
joblib>=1.3.0

# Feature Engineering
feature-engine>=1.6.0,<2.0.0
imbalanced-learn>=0.11.0,<1.0.0

# Visualization
matplotlib>=3.8.0,<4.0.0
seaborn>=0.13.0,<1.0.0
plotly>=5.17.0,<6.0.0

# Model Interpretability & Statistical Analysis
shap>=0.42.0,<1.0.0
scipy>=1.11.0,<2.0.0

# Data Storage
pyarrow>=13.0.0

# Dashboards
streamlit>=1.40.0,<2.0.0
gradio>=4.0.0,<6.0.0

# Image Processing
Pillow>=10.0.0,<11.0.0

# Experiment Tracking & Model Registry
mlflow>=2.9.0,<3.0.0

# ==============================================================================
# Development Dependencies (can be removed before deployment)
# ==============================================================================

# Presentation Generation
python-pptx>=0.6.21,<1.0.0
XlsxWriter>=3.0.0,<4.0.0
lxml>=6.0.0,<7.0.0